clean up and prepare for environmental etc
