+++ lib/cvss_suite/cvss2/cvss2.rb
@@ -9,7 +9,7 @@
 # This work is licensed under the terms of the MIT license.
 # See the LICENSE.md file in the top-level directory.
 
-require_relative '../cvss'
+require_relative '../cvss_v231'
 require_relative 'cvss2_base'
 require_relative 'cvss2_temporal'
 require_relative 'cvss2_environmental'
@@ -17,7 +17,7 @@ require_relative 'cvss2_environmental'
 module CvssSuite
   ##
   # This class represents a CVSS vector in version 2.
-  class Cvss2 < Cvss
+  class Cvss2 < CvssThrough3_1
     ##
     # Returns the Version of the CVSS vector.
     def version
