+++ lib/cvss_suite/cvss31/cvss31.rb
@@ -9,7 +9,7 @@
 # This work is licensed under the terms of the MIT license.
 # See the LICENSE.md file in the top-level directory.
 
-require_relative '../cvss'
+require_relative '../cvss_v231'
 require_relative 'cvss31_base'
 require_relative 'cvss31_temporal'
 require_relative 'cvss31_environmental'
@@ -18,7 +18,7 @@ require_relative '../helpers/cvss31_helper'
 module CvssSuite
   ##
   # This class represents a CVSS vector in version 3.1.
-  class Cvss31 < Cvss
+  class Cvss31 < CvssThrough3_1
     ##
     # Returns the Version of the CVSS vector.
